requirements.txt
opencv-python-headless>=4.10.0.84
numpy>=1.26
pyyaml>=6.0.2
tqdm>=4.66
